% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{assess_all}
\alias{assess_all}
\title{Performs all solution assessments}
\usage{
assess_all(eval_matrix, vert_matrix = NA)
}
\arguments{
\item{eval_matrix}{the evaluation matrix. It is a \eqn{m \times n} matrix with
entries \eqn{e_{ij}}, corresponding to the evaluation of solution \eqn{i}
in the criteria \eqn{j}.}

\item{vert_matrix}{the coordinates of the vertices. It is a \eqn{n \times n} 
matrix with the coordinates of the vertices of the polyhedron of the feasible
weights located at the columns. If \code{NA} then it is assumed that the 
criteria follow a decreasing order of importance, that is, with criterion 
1 more important than criterion 2, and so on.}
}
\value{
A matrix is returned that includes, for each solution, the interval
of scores, whether it is the reference solution or not, and the degrees
of possibility corresponding to the decision maker's attitude.
}
\description{
This function evaluates solutions according to all existing approaches.
That is, both the possibility ones proposed in Torres et al (2021), as
well as the geometric ones.
}
\examples{
eval_mat <- create_eval_matrix_example()
eval_mat \%>\% assess_all()

}
