% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_assess.R
\name{best_vert_prop_compare}
\alias{best_vert_prop_compare}
\title{Best vertices proportion}
\usage{
best_vert_prop_compare(sol_scores, ref_sol_scores)
}
\arguments{
\item{sol_scores}{an n-dimensional vector containing the solution scores at
the vertices.}

\item{ref_sol_scores}{an n-dimensional vector containing the reference
solution scores at the vertices.}
}
\value{
the proportion of the numerical scores that are better than the
reference solution. A value equal to 0 means that the solution has no scores
equal to or better than the reference solution, while a value equal to 1
indicates the opposite. The latter value is achieved by the reference
solution, i.e., if it has been identified as the one with the highest
lower score.
}
\description{
Given a solution scores at the extreme points (vertices), this function
computes the proportion of vertices for which their corresponding scores
are equal or better than the ones from the reference solution. It is
another geometric approach to compare solutions.
}
\examples{
sol_scores <- runif(3)
ref_sol_scores <- runif(3)
best_vert_prop_compare(sol_scores, ref_sol_scores)
}
