% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{blest_rank_correlation}
\alias{blest_rank_correlation}
\title{Blest's measure of rank correlation}
\usage{
blest_rank_correlation(rank1, rank2)
}
\arguments{
\item{rank1}{A permutation of size n of integers in the range [1, n].}

\item{rank2}{A permutation of size n of integers in the range [1, n].}
}
\value{
\eqn{v} correlation coefficient.
}
\description{
This function computes the Blest's measure of rank correlation as defined by 
Blest (2020).
From two ranges of equal size the function returns a scalar in the range 
\eqn{[-1, 1]}, where a value close to -1 means that the ranges differ 
substantially, while one close to 1 means the opposite.
}
\details{
More details can be found in Sałabun, W., Urbaniak, K. (2020). 
A New Coefficient of Rankings Similarity in Decision-Making Problems. 
In: Blest, D.C.: Theory & methods: Rank correlation - an alternative 
measure. Aust. NZ. J. Stat. 42(1), 101–111 (2000). 
https://doi.org/10.1111/1467-842X.00110.
}
\examples{

rank1 <- c(1, 2, 3, 4, 5)
rank2 <- c(3, 1, 2, 4, 5)

blest_rank_correlation(rank1, rank2)

}
