% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic_manipulation.R
\name{compute_intervals}
\alias{compute_intervals}
\title{Scoring interval for each solution}
\usage{
compute_intervals(vertex_scores_matrix)
}
\arguments{
\item{vertex_scores_matrix}{An m by n matrix containing the scores of
each of the m solutions (rows) at each of the n vertices (columns).}
}
\value{
An m by 2 matrix containing the lower bound (LB) and
upper bound (UB) for each of the m solutions (rows).
}
\description{
This function returns, for each solution, its range of possible scores
(intervals).
}
\examples{
# Scoring two solutions on three criteria
m <- matrix(c(1,2,3,4,3,2), byrow = TRUE, nrow = 2)
compute_intervals(m)
}
