% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_with_reference.R
\name{compute_possibility_degree}
\alias{compute_possibility_degree}
\title{Degree of possibility that solution A is superior to solution B.}
\usage{
compute_possibility_degree(A, B, attitude = "neutral")
}
\arguments{
\item{A}{a vector of the score range (interval) for the solution A. The
first element is the lower bound, while the second is the upper bound.}

\item{B}{a vector of the score range (interval) for the solution B and with
the same structure as A.}

\item{attitude}{attitude of the decision-maker. The three possible options
are "neutral" (default), "optimistic", "pessimistic".}
}
\value{
a value between 0 and 1, where a value close to 0 means that
solution A is inferior to B, and a value close to 1, the opposite.
In case A = B, then the degree of possibility is 0.5.
}
\description{
Given the ranges of possible scores (intervals) of two solutions,
this function returns the degree of possibility that solution A is
superior to solution B.
}
\examples{
A <- c(1, 3)
B <- c(2, 4)
# Neutral attitude
compute_possibility_degree(A, B)
# Pessimistic attitude
compute_possibility_degree(A, B, attitude = "pessimistic")
# Optimistic attitude
compute_possibility_degree(A, B, attitude = "optimistic")
}
