% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_sois}
\alias{filter_sois}
\title{Filtering the solution of interest}
\usage{
filter_sois(result_matrix, by, threshold = 0)
}
\arguments{
\item{result_matrix}{a \eqn{m \times p} matrix (\eqn{p \geq 1}) containing an
assessment for each of the \eqn{m} solutions.}

\item{by}{the name of the assessment indicator by which the solutions will
be filtered.}

\item{threshold}{a number corresponding to the threshold.}
}
\value{
a \eqn{k \times p} matrix (\eqn{k \leq m}) with those solutions
that meet the filter condition.
}
\description{
This assistant function filter a result matrix of assessed solutions
according a given threshold. More specifically, it returns those solutions
(rows) with values for the assessment column greater or equal to the
provided threshold.
}
\examples{
create_eval_matrix_example() \%>\%
  score() \%>\%
  intervals() \%>\%
  reference() \%>\%
  poss_assess() \%>\%
  filter_sois(by = "neutral", threshold = 0.0)
}
