% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{generate_criteria_names}
\alias{generate_criteria_names}
\title{Automatic names for criteria}
\usage{
generate_criteria_names(ncrit)
}
\arguments{
\item{ncrit}{The number of criteria}
}
\value{
A character vector of length \code{ncrit}.
}
\description{
It generates names for the criteria consecutively.
}
