% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_polyhedron_vertices}
\alias{generate_polyhedron_vertices}
\title{Vertices of the polyhedron of feasible weights}
\usage{
generate_polyhedron_vertices(n)
}
\arguments{
\item{n}{the number of criteria (dimension of the search space)}
}
\value{
an n by n matrix with the coordinates of each vertex of the
polyhedron in the columns.
}
\description{
Generates an n by n matrix of the vertices of the polyhedron
of the feasible region of weights. This region is induced by the following
set of constraints: wi >= 0 (for all i = 1...n), w1+...+w2 = 1 and w(1) >=...>= w(n).
The order of the columns of the resulting matrix is consistent with
the decision maker's preference over the weights. For example,
the first column corresponds to the most important criterion
(e.g. the first in the order of preference and weight w(1)),
the second column to the second most important criterion, and so on.
This matrix is required to obtain the range of possible scores for each
solution.
}
\examples{
# Generating the coordinates of the three vertices in a 3-criteria problem.
generate_polyhedron_vertices(3)
}
