% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{generate_solutions_names}
\alias{generate_solutions_names}
\title{Automatic names for solutions}
\usage{
generate_solutions_names(nsol)
}
\arguments{
\item{nsol}{The number of solutions.}
}
\value{
A character vector of length \code{nsol}.
}
\description{
It generates names for the solutions consecutively.
}
