% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{generate_weights_names}
\alias{generate_weights_names}
\title{Automatic names for weights}
\usage{
generate_weights_names(ncrit)
}
\arguments{
\item{ncrit}{The number of weights}
}
\value{
A character vector of length \code{ncrit}.
}
\description{
It generates names for the weights consecutively.
}
