% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_based_assess.R
\name{geom_compare}
\alias{geom_compare}
\title{Reference solution comparison}
\usage{
geom_compare(sol_eval, ref_sol_eval, vert_matrix, approach = "volume")
}
\arguments{
\item{sol_eval}{A vector containing the solution evaluation in the extreme
points (vertices of the polyhedron of feasible weights).}

\item{ref_sol_eval}{A vector containing the evaluations of the reference
solution in the extreme points (vertices of the polyhedron of feasible
weights).}

\item{vert_matrix}{The matrix of vertices coordinates.}

\item{approach}{The geometric approach used to compare the solutions. It can
take one of the following character values:
\describe{
  \item{\code{'volume'}}{Normalized volume difference}
  \item{\code{'poss_volume'}}{Normalized volume difference in the range
  \eqn{[0, 1]}.}
  \item{\code{'vert_prop'}}{Proportion of vertices where the solution
  is greater or equal than the reference solution}
}}
}
\value{
A numeric value corresponding to the superiority degree of solution
'sol_eval' over 'ref_sol_eval'.
}
\description{
Compute the degree to which a given solution outperforms the reference
solution. It acts as a function selector for the available geometric
approaches.
}
\examples{
vert_mat <- generate_polyhedron_vertices(3)
ref_sol_eval <- runif(3)
sol_eval <- runif(3)
geom_compare(sol_eval, ref_sol_eval, vert_mat)

}
