% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{geom_identify_sois}
\alias{geom_identify_sois}
\title{Geomtric approach for the identification of the solution of interest}
\usage{
geom_identify_sois(eval_matrix, vert_matrix = NA, by = "volume", threshold = 0)
}
\arguments{
\item{eval_matrix}{the evaluation matrix. It is a m by n matrix with
entries eij where it represents the evaluation provided to solution i
in the criteria j.}

\item{vert_matrix}{the coordinates of the vertices. It is an n by n matrix
with the coordinates of the vertices of the polyhedron corresponding to the w
eights in the columns. If NA then it is assumed that the criteria follow a
decreasing order of importance, that is, with criterion 1 more important
than criterion 2, and so on.}

\item{by}{the geometric assessment approach.
It can take one of the following values: 'volume', 'poss_volume' (possibility
from volume), or 'vert_prop' (vertex proportion).}

\item{threshold}{threshold to identify the solutions of interest.
In particular, a filter is applied on the assessment value of each solution
so that only those with values greater than or equal to threshold will
appear in the returned matrix.}
}
\value{
A matrix is returned that includes, for each solution, the interval
of scores, whether it is the reference solution or not, and the assessments
corresponding to the geometric approach.
}
\description{
Identify the solutions of interest of a multi-criteria decision problem
by applying a geometric assessment approach.
}
\examples{
eval_mat <- create_eval_matrix_example()
eval_mat \%>\% geom_identify_sois()

}
