% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_criteria_prefix}
\alias{get_criteria_prefix}
\title{Name prefix for criteria columns}
\usage{
get_criteria_prefix()
}
\value{
Return the prefix for the automatic names of criteria.
}
\description{
Name prefix for criteria columns
}
