% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_eval_vertex_prefix}
\alias{get_eval_vertex_prefix}
\title{Name prefix for vertices score columns}
\usage{
get_eval_vertex_prefix()
}
\value{
Return the prefix for the automatic names of vertices.
}
\description{
Name prefix for vertices score columns
}
