% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_geom_approaches}
\alias{get_geom_approaches}
\title{Geometric approaches}
\usage{
get_geom_approaches()
}
\value{
a 3-dimensional character vector with the names of
the geometric approaches used to assess the solutions.
}
\description{
Geometric approaches
}
