% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_poss_approaches}
\alias{get_poss_approaches}
\title{Decision-maker attitudes}
\usage{
get_poss_approaches()
}
\value{
a 3-dimensional character vector with the names of
the decision-maker attitudes. These attitudes are used for assessing the
solutions from the possibility approach of Torres et al (2021).
}
\description{
Decision-maker attitudes
}
