% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_prefixed_column_names}
\alias{get_prefixed_column_names}
\title{Generation of the column names for the vertex scores}
\usage{
get_prefixed_column_names(colum_names, prefix = "VE_")
}
\arguments{
\item{colum_names}{a character vector containing several column names}

\item{prefix}{the prefix used to denote the names of the columns 
corresponding to vertex scores}
}
\value{
the column names of the vertices scores
}
\description{
Generation of the column names for the vertex scores
}
