% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_solutions_prefix}
\alias{get_solutions_prefix}
\title{Name prefix for criteria columns}
\usage{
get_solutions_prefix()
}
\value{
Return the prefix for the automatic names of solutions.
}
\description{
Name prefix for criteria columns
}
