% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{get_weights_prefix}
\alias{get_weights_prefix}
\title{Name prefix for weights columns}
\usage{
get_weights_prefix()
}
\value{
Return the prefix for the automatic names of weights.
}
\description{
Name prefix for weights columns
}
