% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Scoring interval for each solution}
\usage{
intervals(eval_at_vert_matrix, append_output = TRUE)
}
\arguments{
\item{eval_at_vert_matrix}{An m by n matrix containing the scores of
each of the m solutions (rows) at each of the n vertices (columns).}

\item{append_output}{A boolean parameter indicating whether (or not) the
results will be append to the input matrix.}
}
\value{
An m by 2 matrix containing the lower bound (LB) and
upper bound (UB) for each of the m solutions (rows).
}
\description{
This function returns, for each solution, its range of possible scores
(intervals).
}
\examples{
# Scoring two solutions on three criteria
m <- matrix(c(1, 2, 3, 4, 3, 2), byrow = TRUE, nrow = 2)
intervals(m)
}
