% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{name_eval_matrix}
\alias{name_eval_matrix}
\title{Automatic names for the rows and columns of the evaluation matrix}
\usage{
name_eval_matrix(eval_matrix)
}
\arguments{
\item{eval_matrix}{the evaluation matrix}
}
\value{
the same matrix with names for its rows and columns.
}
\description{
This function automatically assigns names to the solutions (rows) and
criteria (columns) of an evaluation matrix if they are not already set.
}
\examples{
E <- matrix(runif(15), byrow = TRUE, nrow = 5)
E \%>\% name_eval_matrix()
}
