% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poss_assess.R
\name{neutral_attitude}
\alias{neutral_attitude}
\title{Neutral attitude}
\usage{
neutral_attitude(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
return x
}
\description{
This is an assistant function for computing the possibility degree as
proposed by Torres et al (2021).
}
