% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{one_to_one_swap_sensitivity_analysis}
\alias{one_to_one_swap_sensitivity_analysis}
\title{One-to-one swapping sensitivity analysis}
\usage{
one_to_one_swap_sensitivity_analysis(
  eval_matrix,
  crit_pref,
  rank_crit = "neutral",
  strength = 1
)
}
\arguments{
\item{eval_matrix}{the original problem evaluation matrix.}

\item{crit_pref}{a vector with the preference of criteria such that the 
first element of the vector is the most important criterion, the second, 
the second most important, and so on.}

\item{rank_crit}{the criteria used to rank the solutions (e.g. neutral, optimistic)}

\item{strength}{the distance (size) of the swap between the criteria positions.}
}
\value{
a data frame where each row is a criteria swap and the columns 
contain information about the swap, its associated solution ranking and its 
similarity to the reference ranking.
}
\description{
This function performs a sensitivity analysis based on a one-to-one swapping 
of the preference (order) of the criteria of an MCDM problem. Specifically, 
it generates all possible alternatives (permutations) of a criteria 
preference and evaluates them according to a ranking comparison measure 
(i.e. taking as a reference the ranking obtained by a given order of 
preference of the criteria). It requires as parameters the solution 
evaluation matrix used, the original criterion preference, the criterion 
under which the solutions will be ranked (i.e. neutral, pessimistic, 
optimistic) and the strength of the swapping 
between criteria (i.e. 1, 2,...). As a result, a matrix is returned where the 
first column corresponds to the preference ID and the second column to the 
ranking comparison measure.
}
\examples{
E <- create_eval_matrix_example(20, 5)
P <- c(2,1,5,3,4)
names(P) <- paste0("C", 1:5)[P]
print(P)
sens <- one_to_one_swap_sensitivity_analysis(E, P, "neutral", 1)
print(sens)

}
