% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poss_assess.R
\name{optimistic_attitude}
\alias{optimistic_attitude}
\title{Optimistic attitude}
\usage{
optimistic_attitude(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
return \eqn{x\sqrt{x}}
}
\description{
This is an assistant function for computing the possibility degree as 
proposed by Torres et al (2021).
}
