% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poss_assess.R
\name{pessimistic_attitude}
\alias{pessimistic_attitude}
\title{Pessimistic attitude}
\usage{
pessimistic_attitude(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
return the \eqn{\log(x)}
}
\description{
This is an assistant function for computing the possibility degree as
proposed by Torres et al (2021).
}
