% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stuffs.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Plots the score intervals of the solutions}
\usage{
plot_intervals(interval_matrix, highlight_reference = TRUE)
}
\arguments{
\item{interval_matrix}{A matrix of m rows containing at least three columns:
\code{'LB'}, \code{'UB'} and \code{'REF'}, corresponding to the lower bound
of the interval, the upper bound, and the type of solution (reference
solution = 1, other solution = 0), respectively.}

\item{highlight_reference}{Whether to highlight the reference solution
(default) or not.}
}
\value{
A segment plot with scores on the x-axis and solutions on the y-axis.
}
\description{
This function creates a graph of intervals representing the range of
possible scores that each solution can achieve, i.e., based on the
evaluations by criteria and the order of preference of these criteria.
}
\examples{
create_eval_matrix_example() \%>\%
  assess_all() \%>\%
  plot_intervals()

}
