% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poss_assess.R
\name{poss_assess}
\alias{poss_assess}
\title{Possibility assessment of solutions}
\usage{
poss_assess(
  interval_matrix,
  interval_columns = NA,
  reference_column = NA,
  by = c("neutral"),
  append_output = TRUE
)
}
\arguments{
\item{interval_matrix}{a \eqn{m \times p (p \geq 3)} matrix containing at 
least three columns corresponding to the interval lower bound, interval 
upper bound and the type of solution, respectively. Each row corresponds 
to a solution.}

\item{interval_columns}{a 2-dimensional vector containing the names of the 
matrix column corresponding to lower and upper bounds of the intervals. If 
\code{NA}, then it will be assummed that the names are \code{'LB'} and 
\code{'UB'}.}

\item{reference_column}{the name of column corresponding to the reference
solution.}

\item{by}{a vector indicating the assessment indicators to compute. It can
be one or the combination of the following:
\describe{
  \item{\code{'neutral'}}{Neutral attitude.}
  \item{\code{'optimistic'}}{Optimistic attitude.}
  \item{\code{'pessimistic'}}{Pessimistic attitude.}
  \item{\code{'all'}}{For computing the three above indicators.}
}}

\item{append_output}{A Boolean value indicating whether the output is 
append (default) or not to the input matrix.}
}
\value{
A \eqn{m \times p (p \geq 1)} matrix with the computed assessments 
at the columns
}
\description{
This function computes, for each solution, the possibility degree of being 
superior than the reference solution.
}
\examples{
create_eval_matrix_example(10, 3) \%>\%
score() \%>\%
intervals() \%>\%
reference() \%>\%
poss_assess()

}
