% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{rank_sois}
\alias{rank_sois}
\title{Rank a matrix with assessed solutions}
\usage{
rank_sois(result_matrix, by, decreasing = TRUE)
}
\arguments{
\item{result_matrix}{a \eqn{m \times p} matrix (\eqn{p \geq 1}) containing an
assessment for each of the \eqn{m} solutions.}

\item{by}{the name of the assessment indicator.}

\item{decreasing}{whether or not the solutions are sorted in decreasing
(default) order.}
}
\value{
A matrix with the same dimension as \code{result_matrix} and sorted
according to \code{by} and \code{decreasing}.
}
\description{
This is an assistant function for sorting the solutions according to a
given assessment indicator (ie. \code{'neutral'}, \code{'volume'}, etc.).
}
\examples{
create_eval_matrix_example() \%>\%
  poss_identify_sois(threshold = -1) \%>\%
  rank_sois(by = "neutral")
}
