% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_eval_matrix_csv}
\alias{read_eval_matrix_csv}
\title{Read an evaluation matrix from a csv file}
\usage{
read_eval_matrix_csv(path, solutions_ids = TRUE, criteria_ids = TRUE)
}
\arguments{
\item{path}{the path to the csv file containing the evaluation matrix}

\item{solutions_ids}{a boolean value indicating whether the file includes
identifiers for the solutions in the first column. The default value is TRUE.}

\item{criteria_ids}{a boolean value indicating whether the file includes
identifiers for the criteria in the first row. The default value is TRUE.}
}
\value{
an matrix containing the evaluations
}
\description{
This function helps to read an evaluation matrix from a csv file. The
contents of the file must include, in tabular form, the numerical
evaluations for each solution (rows) in each criterion (column).
If the first column includes an identifier for each solution,
it will be read as a string. The same is true for the criteria.
In case it is indicated, through the corresponding parameters, that the
file does not have these identifiers, then the function will assign names
following an ascending order (e.g. S1, S2,... for the solutions,
and C1, C2,... for the criteria).
}
\examples{
# eval_mat <- read_eval_matrix_csv("eval_mat.csv")

}
