% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{select_sois}
\alias{select_sois}
\title{Selecting the top n solutions of interest}
\usage{
select_sois(result_matrix, by, n)
}
\arguments{
\item{result_matrix}{a \eqn{m \times p} matrix (\eqn{p \geq 1}) containing an
assessment for each of the \eqn{m} solutions.}

\item{by}{the name of the assessment indicator by which the solution will
be selected.}

\item{n}{the number of solutions to be selected.}
}
\value{
a \eqn{n \times p} matrix (\eqn{n \leq m}) with those solutions
with the higher values for the assessment indicator defined in \code{by}.
}
\description{
This is an assistant function to get the top n solutions according to
a given assessment indicator.
}
\examples{
create_eval_matrix_example() \%>\%
  score() \%>\%
  intervals() \%>\%
  reference() \%>\%
  poss_assess() \%>\%
  select_sois(by = "neutral", n = 3)
  
}
