% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vertify}
\alias{vertify}
\title{Polyhedron Vertices from Constraints}
\usage{
vertify(A_matrix, b_vector)
}
\arguments{
\item{A_matrix}{a n by n matrix (coefficient matrix)}

\item{b_vector}{a n-dimensional vector (right hand side)}
}
\value{
an n by n matrix with the vertices of the induced polyhedron
in the columns
}
\description{
This function generates, from a system of constraints on the weights,
the vertices of the polyhedron induced by such a system. As input it
requires an n by n matrix A and an n-dimensional vector b such that
Aw <= b. As a result, an n by n matrix is returned, whose columns are
the vertices of the polyhedron.
}
\examples{
A <- matrix(c(0, 0, -1, -1, 1, 0, 0, -1, 1), byrow = TRUE, nrow = 3)
b <- c(0, 0, 0)
vertify(A, b)

}
