% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_based_assess.R
\name{volume}
\alias{volume}
\title{Volume computation}
\usage{
volume(eval_matrix, vert_matrix, mean_score = FALSE, append_output = TRUE)
}
\arguments{
\item{eval_matrix}{the evaluation matrix. A \eqn{m \times n} matrix.}

\item{vert_matrix}{the vertices matrix (extreme points). A \eqn{n \times n}
matrix containing the coordinates of the vertices of the polyhedron induced
by the preference order provided for the criteria.}

\item{mean_score}{whether the mean score will be computed}

\item{append_output}{whether or not, the output is append to
\code{eval_matrix}}
}
\value{
a \eqn{m \times p} matrix (with \eqn{p \geq 1}) with the volume of
every solution of the eval_matrix.
}
\description{
Computes, for each solution (rows), the volume corresponding to the score
function over the polyhedron of feasible weights.
}
\examples{
E <- create_eval_matrix_example(10, 3)
V <- generate_polyhedron_vertices(3)
E \%>\% volume(vert_matrix = V, mean_score = TRUE)
}
