% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{weighted_rank_measure_of_correlation}
\alias{weighted_rank_measure_of_correlation}
\title{Weighted rank measure of correlation}
\usage{
weighted_rank_measure_of_correlation(rank1, rank2)
}
\arguments{
\item{rank1}{A permutation of size n of integers in the range [1, n].}

\item{rank2}{A permutation of size n of integers in the range [1, n].}
}
\value{
\eqn{w_r} correlation coefficient.
}
\description{
This function computes the weighted rank measure of correlation as defined in 
(da Costa & Soares, 2005).
From two ranges of equal size the function returns a scalar in the range 
\eqn{[-1, 1]}, where a value close to -1 means that the ranges differ 
substantially, while one close to 1 means the opposite.
}
\details{
More details can be found in Pinto da Costa, J., Soares, C.: 
A weighted rank measure of correlation. 
Aust. NZ. J. Stat. 47(4), 515–529 (2005). 
https://doi.org/10.1111/j.1467-842X.2005.00413.x
}
\examples{

rank1 <- c(1, 2, 3, 4, 5)
rank2 <- c(3, 1, 2, 4, 5)

weighted_rank_measure_of_correlation(rank1, rank2)

}
