% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_output_csv}
\alias{write_output_csv}
\title{Writes an output or evaluation matrix to a csv file}
\usage{
write_output_csv(output_matrix, path = "output.csv")
}
\arguments{
\item{output_matrix}{the matrix to be saved}

\item{path}{the path where the file will be saved ("output.csv" is set by
default).}
}
\description{
Writes a solution type matrix (e.g. with solution identifiers as row names)
to a specified csv file.
}
\examples{
# write_output_csv(out_matrix, path="output_matrix.csv")
}
