% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{ws_similarity}
\alias{ws_similarity}
\title{WS Coefficient of Rankings Similarity}
\usage{
ws_similarity(rank1, rank2)
}
\arguments{
\item{rank1}{A permutation of size n of integers in the range [1, n].}

\item{rank2}{A permutation of size n of integers in the range [1, n].}
}
\value{
WS Coefficient.
}
\description{
This function computes the WS coefficient as defined by Sałabun 
and Urbaniak (2020).
From two ranges of equal size the function returns a scalar in the range 
\eqn{[0, 1]}, where a value close to 0 means that the ranges differ 
substantially, while one close to 1 means the opposite.
}
\details{
More details can be found in Sałabun, W., Urbaniak, K. (2020). 
A New Coefficient of Rankings Similarity in Decision-Making Problems. 
In: Krzhizhanovskaya, V., et al. Computational Science – ICCS 2020. ICCS 2020. 
Lecture Notes in Computer Science(), vol 12138. Springer, Cham. 
\url{https://doi.org/10.1007/978-3-030-50417-5_47}.
}
\examples{

rank1 <- c(1, 2, 3, 4, 5)
rank2 <- c(3, 1, 2, 4, 5)

ws_similarity(rank1, rank2)

}
