% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic_manipulation.R
\name{compute_intervals}
\alias{compute_intervals}
\title{Compute the score interval for each solution}
\usage{
compute_intervals(vert_eval_matrix)
}
\arguments{
\item{vert_eval_matrix}{An m by n matrix containing the scores of
each of the m solutions (rows) at each of the n vertices (columns).}
}
\value{
An m by 2 matrix containing the lower bound (LB) and
upper bound (UB) for each of the m solutions (rows).
}
\description{
Compute the score interval for each solution
}
\examples{
m <- matrix(c(1,2,3,4,3,2), byrow = T, nrow = 2)
compute_intervals(m)
}
