% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_polyhedron_vertices}
\alias{generate_polyhedron_vertices}
\title{Vertices of the polyhedron of feasible weights}
\usage{
generate_polyhedron_vertices(ncrit)
}
\arguments{
\item{ncrit}{the number of criteria (dimension of the search space)}
}
\value{
an n by n matrix with the coordinates of each vertex of the
polyhedron in the columns.
}
\description{
Generates an n by n matrix of the vertices of the polyhedron
of the feasible region of weights. This region is induced by the following
set of constraints: wi >= 0 (for all i = 1...n), w1+...+w2 = 1 and
w1 <=...<= wn, where wj is the weight of the jth most important criterion
(according to the order of preference given by the decision maker).
In particular, w1 is the weight of the most important criterion,
while wn is the weight of the least important criterion.
The order of the rows of the resulting matrix is consistent with
this descendant order of preference. For example,
the first row corresponds to the most important criterion
(e.g. the first in the order of preference),
the second column to the second most important criterion, and so on.
The columns of the matrix correspond to the coordinates of the polyhedron.
This matrix is required for computing the range (intervals) of possible
scores for each solution.
}
\examples{
# Generating the coordinates of the three vertices in a 3-criteria problem.
generate_polyhedron_vertices(3)

}
