% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{geom_identify_sois}
\alias{geom_identify_sois}
\title{Geometric approach for the identification of the solution of interest}
\usage{
geom_identify_sois(eval_matrix, vert_matrix = NA, by = "volume", threshold = 0)
}
\arguments{
\item{eval_matrix}{the evaluation matrix. It is a \eqn{m \times n} matrix with
entries \eqn{e_{ij}}, corresponding to the evaluation of solution \eqn{i}
in the criteria \eqn{j}.}

\item{vert_matrix}{the coordinates of the vertices. It is a \eqn{n \times n} 
matrix with the coordinates of the vertices of the polyhedron of the feasible
weights located at the columns. If \code{NA} then it is assumed that the 
criteria follow a decreasing order of importance, that is, with criterion 
1 more important than criterion 2, and so on.}

\item{by}{the geometric assessment approach.
It can take one of the following values: \code{'volume'}, 
\code{'poss_volume'}, or \code{'vert_prop'}.}

\item{threshold}{threshold to identify the solutions of interest.
In particular, a filter is applied on the assessment value of each solution
so that only those with values greater than or equal to threshold will
appear in the returned matrix.}
}
\value{
A matrix is returned that includes, for each solution, the interval
of scores, whether it is the reference solution or not, and the assessments
corresponding to the geometric approach.
}
\description{
Identify the solutions of interest of a multi-criteria decision problem
by applying a geometric assessment approach.
}
\examples{
eval_mat <- create_eval_matrix_example()
eval_mat \%>\% geom_identify_sois()

}
